\name{ambur.hva.gui}
\alias{ambur.hva.gui}

\title{ ambur-hva's Graphical User Interface }
\description{
Provides access to functions via windows and buttons
}
\usage{
ambur.hva.gui()
}

\details{
type 'ambur.gui.hva()' at the command prompt to start the GUI
}
\value{
Tcl Tk GUI with buttons and text boxes to enter in parameters 
for the ambur.hva() function
}

\keyword{ gui }
\keyword{ interface }
